refactor: Move PeriodRepository test utils to fixtures
