+++ nvi-rest/src/test/java/no/sikt/nva/nvi/rest/fetch/FetchNviPeriodsHandlerTest.java
@@ -14,11 +14,11 @@ import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
 import no.sikt.nva.nvi.common.db.PeriodRepository;
+import no.sikt.nva.nvi.common.db.PeriodRepositoryFixtures;
 import no.sikt.nva.nvi.common.service.NviPeriodService;
 import no.sikt.nva.nvi.rest.model.NviPeriodsResponse;
 import no.sikt.nva.nvi.rest.model.UpsertNviPeriodRequest;
 import no.sikt.nva.nvi.test.LocalDynamoTest;
-import no.sikt.nva.nvi.test.TestUtils;
 import no.unit.nva.commons.json.JsonUtils;
 import no.unit.nva.stubs.FakeContext;
 import no.unit.nva.testutils.HandlerRequestBuilder;
@@ -53,8 +53,8 @@ class FetchNviPeriodsHandlerTest extends LocalDynamoTest {
 
   @Test
   void shouldReturnPeriodsWhenUserHasAccessRights() throws IOException {
-    TestUtils.setupPersistedPeriod("2023", periodRepository);
-    TestUtils.setupPersistedPeriod("2024", periodRepository);
+    PeriodRepositoryFixtures.setupPersistedPeriod("2023", periodRepository);
+    PeriodRepositoryFixtures.setupPersistedPeriod("2024", periodRepository);
     handler.handleRequest(createRequestWithAccessRight(AccessRight.MANAGE_NVI), output, context);
     var response = GatewayResponse.fromOutputStream(output, NviPeriodsResponse.class);
